chore: cleaning up some of the `chessie` code
bench: 11371856
