+++ src/board/prng.rs
@@ -30,12 +30,6 @@ impl XoShiRo {
         Self(seeds)
     }
 
-    /// Generates the next pseudo-random number in the sequence.
-    #[inline(always)]
-    pub fn get_next(&mut self) -> u64 {
-        Self::xoshiro(self.0).0
-    }
-
     /// `const` analog of [`XoShiRo::get_next`], returning `(next, Self)`.
     #[inline(always)]
     pub const fn get_next_const(self) -> (u64, Self) {
